requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorlog==6.7.0
docopt==0.6.2
frozenlist==1.3.3
hikari==2.0.0.dev116
hikari-crescent==0.5.2
hikari-miru==3.0.2
idna==3.4
multidict==6.0.4
pipreqs==0.4.11
psycopg2-binary==2.9.5
pytz==2023.3
requests==2.28.2
sigparse==3.0.0
typing_extensions==4.4.0
urllib3==1.26.14
yarg==0.1.9
yarl==1.8.2
